use Wazo release key
And change how packaging is done, i.e. now a gpg key fragment is shipped
in /etc/apt/trusted.gpg.d instead of adding/removing the key to the
local trusted keys.

That said the key now appears twice on most installations (i.e. only the
ISO won't show the key twice) since how we bootstrap the installation
of Wazo by first adding the key manually and not removing it afterwards.
This should be looked at.